Restructure ppr to llvm passes
