+++ mlir/lib/Driver/Pipelines.cpp
@@ -43,6 +43,7 @@
 #include "Gradient/IR/GradientDialect.h"
 #include "Gradient/Transforms/Passes.h"
 #include "Mitigation/Transforms/Passes.h"
+#include "QEC/Transforms/Passes.h"
 #include "Quantum/IR/QuantumDialect.h"
 #include "Quantum/Transforms/Passes.h"
 #include "hlo-extensions/Transforms/Passes.h"
@@ -56,6 +57,8 @@ void createQuantumCompilationStage(OpPassManager &pm)
 {
     pm.addPass(catalyst::quantum::createSplitMultipleTapesPass());
     pm.addNestedPass<ModuleOp>(catalyst::createApplyTransformSequencePass());
+    pm.addPass(catalyst::qec::createLowerQECInitOpsPass());
+    pm.addPass(catalyst::qec::createUnrollConditionalPPRPPMPass());
     pm.addPass(catalyst::createInlineNestedModulePass());
     pm.addPass(catalyst::mitigation::createMitigationLoweringPass());
     pm.addPass(catalyst::quantum::createAdjointLoweringPass());
@@ -141,6 +144,7 @@ void createLLVMDialectLoweringStage(OpPassManager &pm)
     options.useGenericFunctions = true;
     pm.addPass(mlir::createFinalizeMemRefToLLVMConversionPass(options));
     pm.addPass(mlir::createConvertIndexToLLVMPass());
+    pm.addPass(catalyst::qec::createQECConversionPass());
     pm.addPass(catalyst::createCatalystConversionPass());
     pm.addPass(catalyst::quantum::createQuantumConversionPass());
     pm.addPass(catalyst::createAddExceptionHandlingPass());
